Better sync data and time in plots.
Improve the logic that detects the start and, especially, end
dates when reading CSV data.
Also, give up on creating the time labels in JS and go back
to passing them from Python.
It seems like generating them in JS would be better, but it's too
hard to ensure that the JS-generated times correspond exactly with
the python-generated data lists.